fix usersForMentions + api controller
