+++ widgets/EditorWidget.php
@@ -23,7 +23,6 @@ use humhub\libs\Html;
 use humhub\modules\file\libs\FileHelper;
 use humhub\widgets\JsWidget;
 use \Firebase\JWT\JWT;
-use humhub\modules\user\models\User;
 
 /**
  * Description of EditorWidget
@@ -86,10 +85,11 @@ class EditorWidget extends JsWidget
     {
         $module = Yii::$app->getModule('onlyoffice');
 
-        $api = [];
-
-        $api['sendNotifyUrl'] = Url::to(['/onlyoffice/mention/send-notify'], true);
-        $api['makeAnchorUrl'] = Url::to(['/onlyoffice/api/make-anchor'], true);
+        $api = [
+            'sendNotifyUrl' => Url::to(['/onlyoffice/api/send-notify'], true),
+            'makeAnchorUrl' => Url::to(['/onlyoffice/api/make-anchor'], true),
+            'usersForMentionsUrl' => Url::to(['/onlyoffice/api/users-for-mentions'], true),
+        ];
 
         if ($this->file->object_model === cFile::class) {
             $cfile = cFile::findOne($this->file->object_id);
@@ -105,7 +105,6 @@ class EditorWidget extends JsWidget
             'file-info-url' => Url::to(['/onlyoffice/open/get-info', 'guid' => $this->file->guid]),
             'module-configured' => (empty($module->getServerUrl()) ? '0' : '1'),
             'api' => $api,
-            'users-for-mentions' => $this->getUsersForMentions()
         ];
     }
 
@@ -197,22 +196,5 @@ class EditorWidget extends JsWidget
         }
 
         return $config;
-    }
-
-    protected function getUsersForMentions()
-    {
-        $usersForMentions = [];
-        $curUser = Yii::$app->user->getIdentity();
-        $users = User::find()->all();
-
-        foreach ($users as $user) {
-            if ($user->id != $curUser->id && $user->profile->firstname != null && $user->profile->lastname != null && $user->email != null) {
-                array_push($usersForMentions,[
-                    "name" => $user->profile->firstname . " " . $user->profile->lastname,
-                    "email" => $user->email
-                ]);
-            }
-        }
-        return $usersForMentions;
     }
 }
